move test_sso_deployment to shaky test suite
