+++ extension/test/server/right/system_tests_long_right.py
@@ -383,63 +383,7 @@ def test_drop_master_with_load_and_verify():
 
     _test_drop_master_with_load_(client)
 
-@Common.with_custom_setup( Common.setup_1_node_forced_master, Common.basic_teardown )
-def test_sso_deployment():
-    """ the scaling scenario from 1 node to 3 nodes the way A-server does it (eta : 2200 s) """
-    global test_failed
-    test_failed = False 
-    
-    def write_loop (): 
-        Common.iterate_n_times( 10000, 
-                                Common.retrying_set_get_and_delete )
 
-    def large_write_loop (): 
-        Common.iterate_n_times( 280000, 
-                                Common.retrying_set_get_and_delete, 
-                                startSuffix = 1000000 ) 
-
-    write_thr1 = Common.create_and_start_thread ( write_loop )
-    def non_retrying_write_loop (): 
-        Common.iterate_n_times( 10000, 
-                                Common.set_get_and_delete, 
-                                startSuffix = 2000000  )
-    
-    Common.add_node( 1 )
-    cl = _getCluster()
-    cl.setLogLevel("debug")
-    
-    Common.regenerateClientConfig(Common.cluster_id)
-            
-    Common.restart_nodes_wf_sim( 1 )
-    n1 = Common.node_names[1]
-    logging.info("going to start %s", n1)
-    Common.startOne(n1 )
-    
-    Common.create_and_wait_for_thread_list ( [ large_write_loop ] )
-    
-    Common.add_node( 2 )
-    cl = _getCluster()
-    cl.setLogLevel("debug")
-    cl.forceMaster(None )
-    logging.info("2 node config without forced master")
-
-    Common.regenerateClientConfig(Common.cluster_id)
-    
-    Common.restart_nodes_wf_sim( 2 )
-    Common.startOne( Common.node_names[2] )
-    time.sleep( 0.3 )
-    Common.assert_running_nodes ( 3 )
-    
-    write_thr3 = Common.create_and_start_thread ( non_retrying_write_loop )
-
-    write_thr1.join()
-    write_thr3.join()
-    
-    assert_false ( test_failed )
-    
-    Common.assert_running_nodes( 3 )
-    
-    
 @Common.with_custom_setup( Common.setup_3_nodes, Common.basic_teardown )
 def test_3_nodes_2_slaves_down ():
     """ make sure the 'set' operation fails when 2 slaves are down, (eta: 63s) """
